requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
python-decouple==3.8
django-filter==23.3
drf-yasg==1.21.7
markdown==3.5.1